// backend/infrastructure/cdk.out/asset.0bdef5a2cc2d57e707ea7073217ee64c7352ff5dafd182474b36fd78b219df0c/types/api.ts
export interface PaginatedResponse<T> {
  items: T[];
  nextToken?: string;
  total?: number;
}
// ...
export interface BlockedAccount {
  account_id: string;
  username: string;
  handle: string;
  profile_image?: string;
  blocked_at: number;
}

export interface GetBlockListResponse extends PaginatedResponse<BlockedAccount> {}